Rename api.v2 to cloud
